Upgrade to Hyperion 2.0 (#1)
* First steps in migrating site to Hyperion 2.0. Updated DAOs

* ForumViewContext migrated to JPA

* Fixed most obvious compile errors

* Services migrated to Hyperion APIs

* Session/user references updated, tasks rewritten to Hyperion

* Almost all basic pages migrated to new APIs

* Checked admin/group/justice/senate page packages

* Checked event handlers for Hyperion upgrade errors

* Components checked for compile errors

* Compile errors fixed

* Progress in fixing unit tests. Requires twiddling of charts

* Chart stuff fixed, forum queries not quite right yet

* Category queries fixed

* Chancellor pages and tests fixed

* All unit tests green

* Added preconditions to liquibase schema

* Use Argon2 hashes and migrate existing passwords

* Fix logout, fixed forum views

* Fix BPM display

* Style changes

* Reformatting and fixes for date select

* PopulationService user count decreased. Dockerfile and Jenkinsfile added. Bugfix in GraphUtil

+++ src/main/java/com/tysanclan/site/projectewok/entities/dao/ContextBasedForumDAO.java
@@ -17,11 +17,11 @@
  */
 package com.tysanclan.site.projectewok.entities.dao;
 
-import java.util.List;
-
 import com.jeroensteenbeeke.hyperion.data.DomainObject;
 import com.tysanclan.site.projectewok.entities.User;
 import com.tysanclan.site.projectewok.util.forum.ForumViewContext;
+
+import java.util.List;
 
 /**
  * @author Jeroen Steenbeeke
@@ -32,6 +32,6 @@ public interface ContextBasedForumDAO<T extends DomainObject, ContextType extend
 			ForumViewContext context);
 
 	List<T> findByContext(User user, ContextType contextId,
-			ForumViewContext context, long first, long count);
+			ForumViewContext context, int first, int count);
 
 }
